# api-docs.json
{
	"openapi": "3.1.0",
	"info": {
		"description": "SteVe REST API Documentation",
		"license": {
			"name": "GPL-3.0",
			"url": "https://github.com/steve-community/steve/blob/master/LICENSE.txt"
		},
		"title": "SteVe REST API Documentation",
		"version": "3.12.0-SNAPSHOT"
	},
	"servers": [
		{
			"url": "/",
			"description": "Default Server URL"
		}
	],
	"security": [
		{
			"basicAuth": []
		}
	],
...
	"paths": {
...
		"/api/v1/operations/ChangeAvailability": {
			"post": {
				"operationId": "changeAvailability",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangeAvailabilityParams"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OcppOperationResponseAvailabilityStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"404": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Unprocessable Content"
					},
					"500": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"ocpp-operations-controller"
				]
			}
		},
...
		"/api/v1/operations/ClearChargingProfile": {
			"post": {
				"operationId": "clearChargingProfile",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClearChargingProfileParams"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OcppOperationResponseClearChargingProfileStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"404": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Unprocessable Content"
					},
					"500": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"ocpp-operations-controller"
				]
			}
		},
...
		"/api/v1/operations/GetCompositeSchedule": {
			"post": {
				"operationId": "getCompositeSchedule",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetCompositeScheduleParams"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OcppOperationResponseGetCompositeScheduleResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"404": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Unprocessable Content"
					},
					"500": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"ocpp-operations-controller"
				]
			}
		},
		"/api/v1/operations/GetConfiguration": {
			"post": {
				"operationId": "getConfiguration",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetConfigurationParams"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OcppOperationResponseConfigurationKeyValues"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"404": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Unprocessable Content"
					},
					"500": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"ocpp-operations-controller"
				]
			}
		},
...
		"/api/v1/operations/RemoteStartTransaction": {
			"post": {
				"description": "Only 1 charge point can be selected.\n",
				"operationId": "remoteStartTransaction",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RemoteStartTransactionParams"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OcppOperationResponseRemoteStartStopStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"404": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Unprocessable Content"
					},
					"500": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"ocpp-operations-controller"
				]
			}
		},
		"/api/v1/operations/RemoteStopTransaction": {
			"post": {
				"description": "Only 1 charge point can be selected.\n",
				"operationId": "remoteStopTransaction",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RemoteStopTransactionParams"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OcppOperationResponseRemoteStartStopStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"404": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Unprocessable Content"
					},
					"500": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"ocpp-operations-controller"
				]
			}
		},
...
		"/api/v1/operations/SetChargingProfile": {
			"post": {
				"operationId": "setChargingProfile",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SetChargingProfileParams"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OcppOperationResponseChargingProfileStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"404": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Unprocessable Content"
					},
					"500": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"ocpp-operations-controller"
				]
			}
		},
...
		"/api/v1/operations/TriggerMessage": {
			"post": {
				"operationId": "triggerMessage",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TriggerMessageParams"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OcppOperationResponseTriggerMessageStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"404": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Unprocessable Content"
					},
					"500": {
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"ocpp-operations-controller"
				]
			}
		},
...
	},
	"components": {
		"schemas": {
			"ApiErrorResponse": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"status": {
						"type": "integer",
						"format": "int32"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"ChangeAvailabilityParams": {
				"type": "object",
				"properties": {
					"availType": {
						"type": "string",
						"enum": [
							"Inoperative",
							"Operative"
						]
					},
					"chargeBoxIdList": {
						"type": "array",
						"description": "Should contain at least 1 element",
						"items": {
							"type": "string",
							"description": "The identifier of the chargebox (i.e. charging station)"
						},
						"minItems": 1
					},
					"connectorId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					}
				},
				"required": [
					"availType",
					"chargeBoxIdList"
				]
			},
...
			"ClearChargingProfileParams": {
				"type": "object",
				"properties": {
					"chargeBoxIdList": {
						"type": "array",
						"description": "Should contain at least 1 element",
						"items": {
							"type": "string",
							"description": "The identifier of the chargebox (i.e. charging station)"
						},
						"minItems": 1
					},
					"chargingProfilePk": {
						"type": "integer",
						"format": "int32"
					},
					"chargingProfilePurpose": {
						"type": "string",
						"enum": [
							"ChargePointMaxProfile",
							"TxDefaultProfile",
							"TxProfile"
						]
					},
					"connectorId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"filterType": {
						"type": "string",
						"enum": [
							"ChargingProfileId",
							"OtherParameters"
						]
					},
					"stackLevel": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"chargeBoxIdList",
					"filterType"
				]
			},
...
			"GetCompositeScheduleParams": {
				"type": "object",
				"properties": {
					"chargeBoxIdList": {
						"type": "array",
						"description": "Should contain at least 1 element",
						"items": {
							"type": "string",
							"description": "The identifier of the chargebox (i.e. charging station)"
						},
						"minItems": 1
					},
					"chargingRateUnit": {
						"type": "string",
						"enum": [
							"W",
							"A"
						]
					},
					"connectorId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"durationInSeconds": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"chargeBoxIdList",
					"connectorId",
					"durationInSeconds"
				]
			},
...
			"GetConfigurationParams": {
				"type": "object",
				"properties": {
					"chargeBoxIdList": {
						"type": "array",
						"description": "Should contain at least 1 element",
						"items": {
							"type": "string",
							"description": "The identifier of the chargebox (i.e. charging station)"
						},
						"minItems": 1
					},
					"commaSeparatedCustomConfKeys": {
						"type": "string",
						"description": "Comma separated sequence of Custom Configuration Keys"
					},
					"confKeyList": {
						"type": "array",
						"$ref": "#/components/schemas/ConfigurationKeyEnum_Read_Keys",
						"description": "List of Configuration Keys predefined by Ocpp.\nThe documented keys are the superset of all keys defined in Ocpp versions 1.2, 1.5 and 1.6.\nTherefore, not all possible keys apply to all Ocpp versions.\n",
						"items": {
							"type": "string"
						}
					}
				},
				"required": [
					"chargeBoxIdList"
				]
			},
...
			"OcppOperationResponseAvailabilityStatus": {
				"type": "object",
				"properties": {
					"errorResponses": {
						"type": "array",
						"description": "Responses where the station successfully received the request, but responded with a <i>CallError</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/ErrorResponse"
						}
					},
					"exceptions": {
						"type": "array",
						"description": "Exceptions encountered when this platform failed to send the request to the station or failed to process the station's response.\n",
						"items": {
							"$ref": "#/components/schemas/CallException"
						}
					},
					"successResponses": {
						"type": "array",
						"description": "Responses where the station successfully received and processed the request, returning a <i>CallResult</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/SuccessResponseAvailabilityStatus"
						}
					},
					"taskFinished": {
						"type": "boolean",
						"description": "True when all stations have completed communication (either with a <i>CallResult</i>, <i>CallError</i>, or exception) before the platform timeout.\nIf timeout occurs before completion, this will be false and only partial results will be available.\n"
					},
					"taskId": {
						"type": "integer",
						"format": "int32"
					}
				}
			},
...
			"OcppOperationResponseClearChargingProfileStatus": {
				"type": "object",
				"properties": {
					"errorResponses": {
						"type": "array",
						"description": "Responses where the station successfully received the request, but responded with a <i>CallError</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/ErrorResponse"
						}
					},
					"exceptions": {
						"type": "array",
						"description": "Exceptions encountered when this platform failed to send the request to the station or failed to process the station's response.\n",
						"items": {
							"$ref": "#/components/schemas/CallException"
						}
					},
					"successResponses": {
						"type": "array",
						"description": "Responses where the station successfully received and processed the request, returning a <i>CallResult</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/SuccessResponseClearChargingProfileStatus"
						}
					},
					"taskFinished": {
						"type": "boolean",
						"description": "True when all stations have completed communication (either with a <i>CallResult</i>, <i>CallError</i>, or exception) before the platform timeout.\nIf timeout occurs before completion, this will be false and only partial results will be available.\n"
					},
					"taskId": {
						"type": "integer",
						"format": "int32"
					}
				}
			},
			"OcppOperationResponseConfigurationKeyValues": {
				"type": "object",
				"properties": {
					"errorResponses": {
						"type": "array",
						"description": "Responses where the station successfully received the request, but responded with a <i>CallError</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/ErrorResponse"
						}
					},
					"exceptions": {
						"type": "array",
						"description": "Exceptions encountered when this platform failed to send the request to the station or failed to process the station's response.\n",
						"items": {
							"$ref": "#/components/schemas/CallException"
						}
					},
					"successResponses": {
						"type": "array",
						"description": "Responses where the station successfully received and processed the request, returning a <i>CallResult</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/SuccessResponseConfigurationKeyValues"
						}
					},
					"taskFinished": {
						"type": "boolean",
						"description": "True when all stations have completed communication (either with a <i>CallResult</i>, <i>CallError</i>, or exception) before the platform timeout.\nIf timeout occurs before completion, this will be false and only partial results will be available.\n"
					},
					"taskId": {
						"type": "integer",
						"format": "int32"
					}
				}
			},
...
			"OcppOperationResponseGetCompositeScheduleResponse": {
				"type": "object",
				"properties": {
					"errorResponses": {
						"type": "array",
						"description": "Responses where the station successfully received the request, but responded with a <i>CallError</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/ErrorResponse"
						}
					},
					"exceptions": {
						"type": "array",
						"description": "Exceptions encountered when this platform failed to send the request to the station or failed to process the station's response.\n",
						"items": {
							"$ref": "#/components/schemas/CallException"
						}
					},
					"successResponses": {
						"type": "array",
						"description": "Responses where the station successfully received and processed the request, returning a <i>CallResult</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/SuccessResponseGetCompositeScheduleResponse"
						}
					},
					"taskFinished": {
						"type": "boolean",
						"description": "True when all stations have completed communication (either with a <i>CallResult</i>, <i>CallError</i>, or exception) before the platform timeout.\nIf timeout occurs before completion, this will be false and only partial results will be available.\n"
					},
					"taskId": {
						"type": "integer",
						"format": "int32"
					}
				}
			},
...
			"OcppOperationResponseRemoteStartStopStatus": {
				"type": "object",
				"properties": {
					"errorResponses": {
						"type": "array",
						"description": "Responses where the station successfully received the request, but responded with a <i>CallError</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/ErrorResponse"
						}
					},
					"exceptions": {
						"type": "array",
						"description": "Exceptions encountered when this platform failed to send the request to the station or failed to process the station's response.\n",
						"items": {
							"$ref": "#/components/schemas/CallException"
						}
					},
					"successResponses": {
						"type": "array",
						"description": "Responses where the station successfully received and processed the request, returning a <i>CallResult</i> without any communication errors.\n",
						"items": {
							"$ref": "#/components/schemas/SuccessResponseRemoteStartStopStatus"
						}
					},
					"taskFinished": {
						"type": "boolean",
						"description": "True when all stations have completed communication (either with a <i>CallResult</i>, <i>CallError</i>, or exception) before the platform timeout.\nIf timeout occurs before completion, this will be false and only partial results will be available.\n"
					},
					"taskId": {
						"type": "integer",
						"format": "int32"
					}
				}
			},
...
			"RemoteStartTransactionParams": {
				"type": "object",
				"properties": {
					"chargeBoxIdList": {
						"type": "array",
						"description": "Should contain exactly 1 element",
						"items": {
							"type": "string",
							"description": "The identifier of the chargebox (i.e. charging station)"
						},
						"maxItems": 1,
						"minItems": 1
					},
					"chargingProfilePk": {
						"type": "integer",
						"format": "int32",
						"description": "PK of the charging profile to be used for the transaction.\n<code>ChargingProfilePurposeType</code> of the profile must be <code>TX_PROFILE</code>.\n"
					},
					"connectorId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"idTag": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"chargeBoxIdList",
					"idTag"
				]
			},
			"RemoteStopTransactionParams": {
				"type": "object",
				"properties": {
					"chargeBoxIdList": {
						"type": "array",
						"description": "Should contain exactly 1 element",
						"items": {
							"type": "string",
							"description": "The identifier of the chargebox (i.e. charging station)"
						},
						"maxItems": 1,
						"minItems": 1
					},
					"transactionId": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"chargeBoxIdList",
					"transactionId"
				]
			},
...
			"SetChargingProfileParams": {
				"type": "object",
				"properties": {
					"chargeBoxIdList": {
						"type": "array",
						"description": "Should contain at least 1 element",
						"items": {
							"type": "string",
							"description": "The identifier of the chargebox (i.e. charging station)"
						},
						"minItems": 1
					},
					"chargingProfilePk": {
						"type": "integer",
						"format": "int32"
					},
					"connectorId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"transactionId": {
						"type": "integer",
						"format": "int32",
						"description": "Apply the profile to the transaction with this ID.\n<code>ChargingProfilePurposeType</code> of the profile must be <code>TX_PROFILE</code>.\n"
					}
				},
				"required": [
					"chargeBoxIdList",
					"chargingProfilePk",
					"connectorId"
				]
			},
...
			"TriggerMessageParams": {
				"type": "object",
				"properties": {
					"chargeBoxIdList": {
						"type": "array",
						"description": "Should contain at least 1 element",
						"items": {
							"type": "string",
							"description": "The identifier of the chargebox (i.e. charging station)"
						},
						"minItems": 1
					},
					"connectorId": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					},
					"triggerMessage": {
						"type": "string",
						"enum": [
							"BootNotification",
							"DiagnosticsStatusNotification",
							"FirmwareStatusNotification",
							"Heartbeat",
							"MeterValues",
							"StatusNotification"
						]
					}
				},
				"required": [
					"chargeBoxIdList",
					"triggerMessage"
				]
			},
...
		},
		"securitySchemes": {
			"basicAuth": {
				"name": "basicAuth",
				"scheme": "basic",
				"type": "http"
			}
		}
	}
}